requirements.txt
arrow==1.3.0
Flask==3.1.0
flask_limiter==3.12
Flask_Login==0.6.3
python-dotenv==1.0.1
